+++ .eslintrc.js
@@ -244,7 +244,6 @@ module.exports = {
     DTRACE_HTTP_SERVER_REQUEST: false,
     DTRACE_HTTP_SERVER_RESPONSE: false,
     DTRACE_NET_SERVER_CONNECTION: false,
-    DTRACE_NET_STREAM_END: false,
-    internalBinding: false,
+    DTRACE_NET_STREAM_END: false
   },
 };
