+++ lib/assert.js
@@ -36,7 +36,7 @@ const { openSync, closeSync, readSync } = require('fs');
 const { parseExpressionAt } = require('internal/deps/acorn/dist/acorn');
 const { inspect } = require('util');
 const { EOL } = require('os');
-const nativeModule = require('native_module');
+const { NativeModule } = require('internal/bootstrap_loaders');
 
 // Escape control characters but not \n and \t to keep the line breaks and
 // indentation intact.
@@ -163,7 +163,7 @@ function getErrMessage(call) {
   }
 
   // Skip Node.js modules!
-  if (filename.endsWith('.js') && nativeModule.exists(filename.slice(0, -3))) {
+  if (filename.endsWith('.js') && NativeModule.exists(filename.slice(0, -3))) {
     errorCache.set(identifier, undefined);
     return;
   }
