+++ lib/domain.js
@@ -34,6 +34,7 @@ const {
   ERR_UNHANDLED_ERROR
 } = require('internal/errors').codes;
 const { createHook } = require('async_hooks');
+const { internalBinding } = require('internal/bootstrap_loaders');
 
 // overwrite process.domain with a getter/setter that will allow for more
 // effective optimizations
