+++ lib/internal/bootstrap_node.js
@@ -4,11 +4,16 @@
 // responsible for bootstrapping the node.js core. As special caution is given
 // to the performance of the startup process, many dependencies are invoked
 // lazily.
+//
+// Before this file is run, lib/internal/bootstrap_loaders.js gets run first
+// to bootstrap the internal binding and module loaders, including
+// process.binding(), process._linkedBinding(), internalBinding() and
+// NativeModule. And then { internalBinding, NativeModule } will be passed
+// into this bootstrapper to bootstrap Node.js core.
 
 'use strict';
 
-(function(process) {
-  let internalBinding;
+(function bootstrapNodeJSCore(process, { NativeModule }) {
   const exceptionHandlerState = { captureFn: null };
 
   function startup() {
@@ -268,54 +273,6 @@
       }
     }
     perf.markMilestone(NODE_PERFORMANCE_MILESTONE_BOOTSTRAP_COMPLETE);
-  }
-
-  const moduleLoadList = [];
-  Object.defineProperty(process, 'moduleLoadList', {
-    value: moduleLoadList,
-    configurable: true,
-    enumerable: true,
-    writable: false
-  });
-
-  {
-    const bindingObj = Object.create(null);
-
-    const getBinding = process.binding;
-    process.binding = function binding(module) {
-      module = String(module);
-      let mod = bindingObj[module];
-      if (typeof mod !== 'object') {
-        mod = bindingObj[module] = getBinding(module);
-        moduleLoadList.push(`Binding ${module}`);
-      }
-      return mod;
-    };
-
-    const getLinkedBinding = process._linkedBinding;
-    process._linkedBinding = function _linkedBinding(module) {
-      module = String(module);
-      let mod = bindingObj[module];
-      if (typeof mod !== 'object')
-        mod = bindingObj[module] = getLinkedBinding(module);
-      return mod;
-    };
-  }
-
-  {
-    const bindingObj = Object.create(null);
-
-    const getInternalBinding = process._internalBinding;
-    delete process._internalBinding;
-
-    internalBinding = function internalBinding(module) {
-      let mod = bindingObj[module];
-      if (typeof mod !== 'object') {
-        mod = bindingObj[module] = getInternalBinding(module);
-        moduleLoadList.push(`Internal Binding ${module}`);
-      }
-      return mod;
-    };
   }
 
   function setupProcessObject() {
@@ -601,134 +558,6 @@
     // compile the script, this will throw if it fails
     new vm.Script(source, { displayErrors: true, filename });
   }
-
-  // Below you find a minimal module system, which is used to load the node
-  // core modules found in lib/*.js. All core modules are compiled into the
-  // node binary, so they can be loaded faster.
-
-  const ContextifyScript = process.binding('contextify').ContextifyScript;
-  function runInThisContext(code, options) {
-    const script = new ContextifyScript(code, options);
-    return script.runInThisContext();
-  }
-
-  function NativeModule(id) {
-    this.filename = `${id}.js`;
-    this.id = id;
-    this.exports = {};
-    this.loaded = false;
-    this.loading = false;
-  }
-
-  NativeModule._source = process.binding('natives');
-  NativeModule._cache = {};
-
-  const config = process.binding('config');
-
-  NativeModule.require = function(id) {
-    if (id === 'native_module') {
-      return NativeModule;
-    }
-
-    const cached = NativeModule.getCached(id);
-    if (cached && (cached.loaded || cached.loading)) {
-      return cached.exports;
-    }
-
-    if (!NativeModule.exists(id)) {
-      // Model the error off the internal/errors.js model, but
-      // do not use that module given that it could actually be
-      // the one causing the error if there's a bug in Node.js
-      const err = new Error(`No such built-in module: ${id}`);
-      err.code = 'ERR_UNKNOWN_BUILTIN_MODULE';
-      err.name = 'Error [ERR_UNKNOWN_BUILTIN_MODULE]';
-      throw err;
-    }
-
-    moduleLoadList.push(`NativeModule ${id}`);
-
-    const nativeModule = new NativeModule(id);
-
-    nativeModule.cache();
-    nativeModule.compile();
-
-    return nativeModule.exports;
-  };
-
-  NativeModule.requireForDeps = function(id) {
-    if (!NativeModule.exists(id) ||
-        // TODO(TimothyGu): remove when DEP0084 reaches end of life.
-        id.startsWith('node-inspect/') ||
-        id.startsWith('v8/')) {
-      id = `internal/deps/${id}`;
-    }
-    return NativeModule.require(id);
-  };
-
-  NativeModule.getCached = function(id) {
-    return NativeModule._cache[id];
-  };
-
-  NativeModule.exists = function(id) {
-    return NativeModule._source.hasOwnProperty(id);
-  };
-
-  if (config.exposeInternals) {
-    NativeModule.nonInternalExists = NativeModule.exists;
-
-    NativeModule.isInternal = function(id) {
-      return false;
-    };
-  } else {
-    NativeModule.nonInternalExists = function(id) {
-      return NativeModule.exists(id) && !NativeModule.isInternal(id);
-    };
-
-    NativeModule.isInternal = function(id) {
-      return id.startsWith('internal/');
-    };
-  }
-
-
-  NativeModule.getSource = function(id) {
-    return NativeModule._source[id];
-  };
-
-  NativeModule.wrap = function(script) {
-    return NativeModule.wrapper[0] + script + NativeModule.wrapper[1];
-  };
-
-  NativeModule.wrapper = [
-    '(function (exports, require, module, internalBinding, process) {',
-    '\n});'
-  ];
-
-  NativeModule.prototype.compile = function() {
-    let source = NativeModule.getSource(this.id);
-    source = NativeModule.wrap(source);
-
-    this.loading = true;
-
-    try {
-      const fn = runInThisContext(source, {
-        filename: this.filename,
-        lineOffset: 0,
-        displayErrors: true
-      });
-      const requireFn = this.id.startsWith('internal/deps/') ?
-        NativeModule.requireForDeps :
-        NativeModule.require;
-      fn(this.exports, requireFn, this, internalBinding, process);
-
-      this.loaded = true;
-    } finally {
-      this.loading = false;
-    }
-  };
-
-  NativeModule.prototype.cache = function() {
-    NativeModule._cache[this.id] = this;
-  };
 
   startup();
 });
