+++ lib/internal/loader/CreateDynamicModule.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { internalBinding } = require('internal/bootstrap_loaders');
 const { ModuleWrap } = internalBinding('module_wrap');
 const debug = require('util').debuglog('esm');
 const ArrayJoin = Function.call.bind(Array.prototype.join);
