+++ lib/internal/loader/DefaultResolve.js
@@ -3,7 +3,7 @@
 const { URL } = require('url');
 const CJSmodule = require('module');
 const internalFS = require('internal/fs');
-const NativeModule = require('native_module');
+const { NativeModule, internalBinding } = require('internal/bootstrap_loaders');
 const { extname } = require('path');
 const { realpathSync } = require('fs');
 const preserveSymlinks = !!process.binding('config').preserveSymlinks;
