+++ lib/internal/loader/ModuleJob.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { internalBinding } = require('internal/bootstrap_loaders');
 const { ModuleWrap } = internalBinding('module_wrap');
 const { SafeSet, SafePromise } = require('internal/safe_globals');
 const { decorateErrorStack } = require('internal/util');
