+++ lib/internal/loader/Translators.js
@@ -1,7 +1,7 @@
 'use strict';
 
+const { NativeModule, internalBinding } = require('internal/bootstrap_loaders');
 const { ModuleWrap } = internalBinding('module_wrap');
-const NativeModule = require('native_module');
 const internalCJSModule = require('internal/module');
 const CJSModule = require('module');
 const internalURLModule = require('internal/url');
