+++ lib/internal/process/modules.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { internalBinding } = require('internal/bootstrap_loaders');
 const {
   setImportModuleDynamicallyCallback,
   setInitializeImportMetaObjectCallback
