+++ lib/internal/test/binding.js
@@ -8,6 +8,7 @@ process.emitWarning(
 // These exports should be scoped as specifically as possible
 // to avoid exposing APIs because even with that warning and
 // this file being internal people will still try to abuse it.
+const { internalBinding } = require('internal/bootstrap_loaders');
 module.exports = {
   ModuleWrap: internalBinding('module_wrap').ModuleWrap,
 };
