+++ lib/internal/vm/Module.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { internalBinding } = require('internal/bootstrap_loaders');
 const { emitExperimentalWarning } = require('internal/util');
 const { URL } = require('internal/url');
 const { kParsingContext, isContext } = process.binding('contextify');
