+++ lib/module.js
@@ -21,7 +21,7 @@
 
 'use strict';
 
-const NativeModule = require('native_module');
+const { NativeModule } = require('internal/bootstrap_loaders');
 const util = require('util');
 const { decorateErrorStack } = require('internal/util');
 const { getURLFromFilePath } = require('internal/url');
