+++ lib/string_decoder.js
@@ -22,6 +22,7 @@
 'use strict';
 
 const { Buffer } = require('buffer');
+const { internalBinding } = require('internal/bootstrap_loaders');
 const {
   kIncompleteCharactersStart,
   kIncompleteCharactersEnd,
