+++ src/node_javascript.h
@@ -29,7 +29,8 @@
 namespace node {
 
 void DefineJavaScript(Environment* env, v8::Local<v8::Object> target);
-v8::Local<v8::String> MainSource(Environment* env);
+v8::Local<v8::String> LoadersBootstrapperSource(Environment* env);
+v8::Local<v8::String> NodeBootstrapperSource(Environment* env);
 
 }  // namespace node
