+++ test/message/error_exit.out
@@ -12,4 +12,4 @@ AssertionError [ERR_ASSERTION]: 1 strictEqual 2
     at Function.Module._load (module.js:*:*)
     at Function.Module.runMain (module.js:*:*)
     at startup (bootstrap_node.js:*:*)
-    at bootstrap_node.js:*:*
+    at bootstrapNodeJSCore (bootstrap_node.js:*:*)
